Trying to make wifi work without timeouts
Watchdog is currently timing out but can fix that by setting the task priority to same as idle = 0. Might do that but wanted to try a higher priority to see if it makes the wifi work faster or connect better. Spoiler alert: it doesn't.

So set priority back to 0 and keep trying to see if readBytes can be faster? It likes to hang.